Adds full kimlik validation
Full kimlik validation is based on the following facts:

- Composed of 11 digits.

- First digit is not 0.

- When the (sum of 2nd, 4th, 6th, 8th elements) is substraced from (the
multiplication of 7 and (1st 3rd 5th 7th and 9th elements)) and the
resultan is modulated by 10 the final result is the 10th element.

- The sum of first 10 elements modulated by 10 yields the 11th element.

This is not factored in the most optimal form, just quickly layed out
the algorithm; it could be refactored.